eventFeed widget renamed
